requirements.txt
tensorflow
flask
keras
imutils
cv2
werkzeug.utils